fixes for Visual Studio

+++ arangod/Agency/Supervision.cpp
@@ -462,7 +462,7 @@ void Supervision::run() {
           }
         }
       }
-      _cv.wait(1000000 * _frequency);
+      _cv.wait(static_cast<uint64_t>(1000000 * _frequency));
     }
   }
   if (shutdown) {
